--- PETINA/Pruning/__init__.py
@@ -1,12 +1,0 @@
-
-from .pruning import (
-    applyPruning,
-    applyPruningAdaptive,
-    applyPruningDP,
-)
-
-__all__ = [
-    "applyPruning",
-    "applyPruningAdaptive",
-    "applyPruningDP",  
-]
