--- PETINA/Sketching/__init__.py
@@ -1,14 +1,0 @@
-from .sketch import (
-    centralized_count_mean_sketch,
-    generate_hash_funcs,
-    Client_PETINA_CMS,
-    Server_PETINA_CMS,
-    applyCountSketch
-)
-__all__ = [
-    "centralized_count_mean_sketch",
-    "generate_hash_funcs",
-    "Client_PETINA_CMS",
-    "Server_PETINA_CMS",
-    "applyCountSketch"
-]
